organize into stylesheet
