+++ components/DataTableComponent/TableRow.tsx
@@ -1,36 +1,12 @@
 import { css } from '@emotion/react';
 import { Row, flexRender } from '@tanstack/react-table';
 import { useState } from 'react';
+import { linkStyle, rowStyle, tdStyle } from './styles';
 
 type Props<T> = {
 	row: Row<T>;
 	index: number;
 };
-
-const tdStyle = css`
-	padding: 12px;
-	border-bottom: 1px solid #eaeaea;
-	max-width: 30vw;
-	white-space: pre-wrap;
-	overflow-wrap: break-word;
-	word-break: break-word;
-	vertical-align: top;
-`;
-
-const rowStyle = (index: number) => css`
-	background-color: ${index % 2 === 0 ? '' : '#F5F7F8'};
-`;
-
-const linkStyle = css`
-	color: #0b75a2;
-	font-size: 12px;
-	cursor: pointer;
-	margin-left: 6px;
-
-	&:hover {
-		text-decoration: underline;
-	}
-`;
 
 const TableRow = <T,>({ row, index }: Props<T>) => {
 	const [expandedCells, setExpandedCells] = useState<Record<string, boolean>>({});
